refactor: move credential/serializer.py to credential/serializer_v1.py to ease future removal
Relates to JIRA: DISCOVERY-797
