+++ quipucords/api/credential/serializer_v1.py
